+++ solutions/cron/.env.example
@@ -1,0 +1,4 @@
+EDGE_CONFIG_ID=
+EDGE_CONFIG=
+TEAM_ID_VERCEL=
+ACCESS_TOKEN_VERCEL=
