+++ solutions/cron/README.md
@@ -1,11 +1,21 @@
 ---
-marketplace: false
-reason: https://github.com/lfades/static-tweet is very similar to this one
+name: Vercel Cron Job Example
+slug: vercel-cron
+description: A Next.js app that uses Vercel Cron Jobs to update data at different intervals.
+framework: Next.js
+useCase:
+  - Cron
+  - Edge Functions
+  - Edge Config
+css: Tailwind
+demoUrl: https://cron.vercel.app/
+relatedTemplates:
+  - hacker-news-slack-bot
 ---
 
 # Vercel Cron Job Example
 
-This example shows you how you can use Vercel Cron Jobs to update data at different intervals
+A Next.js app that uses [Vercel Cron Jobs](https://vercel.com/docs/cron-jobs) to update data at different intervals.
 
 ## Demo
 
@@ -15,11 +25,26 @@ https://cron.vercel.app/
 
 You can choose from one of the following two methods to use this repository:
 
+### 1. Set Up Environment variables
+
+You'll need to complete the following steps to get the required environment variables:
+
+1. Follow [these steps](https://vercel.com/docs/concepts/edge-network/edge-config/get-started#create-an-edge-config-in-your-account) to configure your Edge Config store. This should give you both the `EDGE_CONFIG_ID` and the `EDGE_CONFIG_URL` variables.
+
+2. Get your remaining environment variables:
+
+- `TEAM_ID_VERCEL`: The Vercel Team ID you created the Edge Config store under. If you're creating this under your personal account, you can omit this variable.
+- `ACCESS_TOKEN_VERCEL`: Your Vercel Personal Access Token scoped to your appropriate Vercel Team.
+
+### 2. Deploy to Vercel
+
 **One-Click Deploy**
 
 Deploy the example using [Vercel](https://vercel.com?utm_source=github&utm_medium=readme&utm_campaign=examples-repo):
 
 [![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/new/clone?repository-url=https://github.com/vercel/examples/tree/main/solutions/cron&project-name=cron&repository-name=cron)
+
+Don't forget to set the required environment variables that you got from the previous step.
 
 **Clone and Deploy**
